Fixed code example - missing necessary constructor for example
Without the constructor method, the example in the "Persisting new objects" section does not work, it only creates a blank person node.  This example seems to work with a fresh install because these full example code is already included and autoloaded.  But if you try to follow this example as-is in a new namespace, then you run into this issue.